% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_colours.R
\name{luke_colours}
\alias{luke_colours}
\title{Luke Colours}
\usage{
luke_colours(format = "web", naming = "informal", ...)
}
\arguments{
\item{format}{Return the colors in format: web, rgb, or hsv}

\item{naming}{Naming of the colours: informal (unofficial) or PMS (for printing)}

\item{...}{ignored}
}
\description{
Return the official colour palette
}
