% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_logo.R
\name{luke_logo}
\alias{luke_logo}
\title{Luke Logo in Colors}
\usage{
luke_logo(
  bg = "#ffffff",
  bubble = "#00b5e2",
  luke = "#ff8200",
  text = "#3d3b3c",
  bg_opacity = 1,
  lang = "fi",
  file = "luke_logo.png",
  ...
)
}
\arguments{
\item{bg}{color for background}

\item{bubble}{color for bubble}

\item{luke}{color for "LUKE"}

\item{text}{color for the text under "LUKE"}

\item{bg_opacity}{opacity of bg, 0-1.}

\item{lang}{language for the text, "fi", "en" eller "se".}

\item{file}{optional file. Default is "luke_logo.png" in the working directory. See details.}

\item{...}{passed on to rsvg_*, mainly 'width' and 'height' arguments for png's.}
}
\description{
Luke's logo consists of background, the text bubble, "Luke", and full name depending on language. Save it as an image in freely chosen colors.
}
\details{
give all colors in web-format, eg. "#ffffff" for white. See luke_colors().

'file' should end in desired format, e.g. "my_luke_logo.png" to have a png-image as output. 
Supported formats are 'pdf', 'svg' and 'png'.
}
