% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sql_file.R
\name{run_sql_file}
\alias{run_sql_file}
\title{Run SQL From a File}
\usage{
run_sql_file(con, sql_file)
}
\arguments{
\item{con}{A database connection}

\item{sql_file}{A SQL file to run on the database}
}
\value{
tibble
}
\description{
Run SQL From a File
}
\examples{
\dontrun{
pats <- run_sql_file(my_conn,
                     here::here("src/get_patients.sql"))}
}
