% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sql_string.R
\name{run_sql_string}
\alias{run_sql_string}
\title{Run a SQL string on a database}
\usage{
run_sql_string(con, sql_string)
}
\arguments{
\item{con}{A database connection}

\item{sql_string}{A sql string to run on the database}
}
\value{
a tibble
}
\description{
Run a SQL string on a database
}
