% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_ggplot.R
\name{scale_color_luke_c}
\alias{scale_color_luke_c}
\title{GG Continuous Scale}
\usage{
scale_color_luke_c(
  ...,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour",
  ncolours = 8
)
}
\arguments{
\item{...}{passed on to luke_pal}
}
\description{
GG Continuous Scale
}
\details{
luke_pal is used for creating 6 colors for interpolation.
}
