% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_ggplot.R
\name{scale_color_luke_d}
\alias{scale_color_luke_d}
\title{GG Discrete Scale}
\usage{
scale_color_luke_d(..., aesthetics = "colour")
}
\arguments{
\item{...}{passed on to luke_pal}

\item{aesthetic}{defaults to "colour"}
}
\description{
GG Discrete Scale
}
