% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_colours.R
\name{luke_colours}
\alias{luke_colours}
\title{Luke Colours}
\usage{
luke_colours(type = "web", names = "informal", ...)
}
\arguments{
\item{type}{web, rgb, or hsv}

\item{names}{informal or PMS}

\item{...}{ignored}
}
\description{
Luke Colours
}
