% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_fonts.R
\name{luke_fonts}
\alias{luke_fonts}
\title{Luke Fonts}
\usage{
luke_fonts(approx = TRUE, ...)
}
\arguments{
\item{approx}{Return a free approximation instead of the official commercial font?}

\item{...}{ignored}
}
\description{
Return font info on the luke's fonts.
}
\details{
The official font family is "Accura", which is a commercial font published by MyFonts.com. A close approximation
}
