% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_presentation.R
\name{luke_presentation}
\alias{luke_presentation}
\title{Templates for Presentations}
\usage{
luke_presentation(path = ".", type = "pp", lang = "en")
}
\arguments{
\item{path}{path to location where to create template}

\item{type}{Powerpoint ('pp') or LaTeX Beamer ('beamer')}

\item{lang}{'fi', 'en' or 'se'}
}
\description{
Create a presentation template in a desired location.
}
