% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_colours.R
\name{scale_color_luke_d}
\alias{scale_color_luke_d}
\title{GG Discrete Scale}
\usage{
scale_color_luke_d(
  ...,
  alpha = 1,
  begin = 0.25,
  end = 0.78,
  direction = 1,
  aesthetics = "colour"
)
}
\description{
GG Discrete Scale
}
