% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luke_ggplot.R
\name{scale_color_luke_d}
\alias{scale_color_luke_d}
\title{GG Discrete Scale}
\usage{
scale_color_luke_d(
  ...,
  alpha = 1,
  begin = 0.3,
  end = 0.7,
  direction = 1,
  option = "A",
  aesthetics = "colour",
  iwh = FALSE
)
}
\description{
GG Discrete Scale
}
